add cli_mfg_silent_exit label for exiting without trailing SUCCESS/FAILED message
This patch adds a label toward the end of cli_mfg() that will skip printing
of the status message. It also consolidates several exit points.

This has the side-effect of fixing a bug that was introduced with the large
integration, where the --get-size command displays an unwanted status message
after displaying the capacity of the chip.

BUG=
TEST=tested on mario

Review URL: http://codereview.chromium.org/5733005