add _POSIX_C_SOURCE test for portability
This patch is intended to help determine if the ipc lock can be used on the
machine the sources are compiled for.

The Makefile will run a script that generates a C file to check if the
_POSIX_C_SOURCE macro is defined and is greater than 200112 (POSIX.1-2001),
which the semctl() man page indicates is the minimum required version.

If all looks good, the locking code will be compiled and USE_BIG_LOCK
will be added to the FEATURE_CFLAGS.

BUG=
TEST=Tested by locally compiling and checking for the macro definition.

Review URL: http://codereview.chromium.org/6275014