ichspi.c: lower down the ICH7 chipset CDS timeout from 60s to 1ms.
The loooong time is not necessary actually. Change to 30 secs for a full chip erase. Also exit early if an error happens.

The long time would make flashrom look like hung for forever especially when
probing hundreds of flashes chip on ICH7M. Remove this can get failed message
less than 1 sec in total.

We were good on EEEPC and Lenovo S10-3t because we never launched 2 or more flashrom in boot time (one hung for forever is fine as long as not in boot path). But now, activate_date and dump_vpd_log hit this issue.

Change-Id: I5a13f80673b452159c8dcf3fa95ea2d5a4944b9c

R=reinauer@chromium.org,dhendrix@chromium.org
BUG=chromium-os:13092
TEST=Tested all following combinations: {on Lenovo S10-t3, EEEPC, Alex and Mario} for {BIOS and EC} flashes with {-r / -w} commands.

Review URL: http://codereview.chromium.org/6698020