Don't exit update mode after writing status bits.
If the write protected is enabled (which means range bits are not zero),
flashrom would try to clean range bits to zero before reading/writing.
In the flashrom write status function, it exits EC update mode which makes
the EC starts running on the SPI (contract to running on the internal RAM
in update mode). This slows down the flashrom reading because EC is reading
SPI too.

BUS=chrome-os-partner:3668
TEST=Run on Z* machine.
% flashrom --wp-range 0 0x40000; flashrom --wp-enable --wp-status
% time flashrom -r /tmp/ec  # around 6 secs

Change-Id: Ib05a915baddc2a7928d8ad3864ff16b621254eef
Reviewed-on: http://gerrit.chromium.org/gerrit/845
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Yung-Chieh Lo <yjlou@chromium.org>
Tested-by: Yung-Chieh Lo <yjlou@chromium.org>