Add EN25Q series support
This adds support for the following Eon chips: EN25Q40, EN25Q80, EN25Q32A/B,
EN25Q64, and EN25Q128.

It also adds a note that the EN25Q16 is an "evil twin" of the EN25D16.

Datasheets available from Eon @ http://www.eonssi.com/products/products.aspx

This CL is the same as http://gerrit.chromium.org/gerrit/#change,1474, but
without the undesired 1-line change to the Makefile sneaking in and breaking
the build.

BUG=None
TEST=Not at all (locally compiled, but no real hardware)

Change-Id: Id358d7098660356e81702a0a8cf5fb2d726051f1
Reviewed-on: http://gerrit.chromium.org/gerrit/1494
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>