The return value of write protection functions should be chained.
If we specify multiple wp functions in one command, the flashrom exit value
will be the return value of the last called functions. For example:

  % flashrom --wp-disable --wp-status

The flashrom always returns 0 (success) even if the wp_disable is failed.

So, the function return value are chained with OR operation (consider
different return value scheme).

BUG=no bug id
TEST=Run the following commands:
Make sure the hardware WP pin is asserted.
% flashrom --wp-range 0 0x400000 --wp-enable --wp-status ; echo $?
Expect the return value is 0. Then,

% flashrom --wp-disable --wp-status; echo $?
Expect 1 (failed).

Change-Id: I72ed5479b9906bc0913111f845be335f25a4d286
Reviewed-on: http://gerrit.chromium.org/gerrit/2606
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>