Returns error if user specifies -i, but no layout info is available.
Actually we addressed this issue during the code review (line 256 of
http://codereview.chromium.org/6625011/diff/7001/layout.c). But somehow
David committed the code with returning 0, I didn't realize this is buggy.
So, here is the patch.

This patch would check if layout info (either from layout file or fmap on BIOS)
is available when processing included arugments. If not available, return
failed.

BUG=chromeos-paternet:2418
TEST=Tested on K*n (the flash image without fmap).

Change-Id: I2898850f2a9ead444c4a73b3725e2d8050434c04
Reviewed-on: http://gerrit.chromium.org/gerrit/2663
Tested-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>
Reviewed-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>