[2/3] add ICH/PCH flash descriptor decoding
based on the work of Matthias 'mazzoo' Wenzel this patch adds...
 - pretty printing of those ICH/PCH flash descriptors that are cached/mapped
   by the chipset (and which are therefore reachable via FDOC/FDOD
   registers)
 - a new tool to the util directory that is able to decode all flash
   descriptors stored in a flash dump file.

it also adds macros and pretty printing for "Vendor Specific
Component Capabilities" registers:
there are two of them: lower and upper. they describe the properties
of the address space divided by FPBA (which allows to use multiple
flash chips or partitions with different properties). the
prettyprinting function is used by the descriptor code too: the
properties of flash chips supported by the ME are stored in the same
format in a descriptor table. this is also the reason why the pp
function and the register bit macros are not defined in ichspi.c but
ich_descriptors.h

Change-Id: Ibe312573daaf469ece63f7e4e4d23c4a430a67c9
Reviewed-on: http://gerrit.chromium.org/gerrit/3984
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>