Add the --diff <file> longopt.
Before Flashrom calculates which regions to erase and write, it obtains a
reference image by reading content from ROM.

This patch allows expert users to provide a reference image as a file, e.g.
a golden image. This allows the user to take advantage of existing partial write
logic which will "diff" the reference image with the "-w" argument. Also, it
avoids the overhead of reading the original reference image from ROM.

Simply add "--diff <file>" to the command.

(Credit goes to Anton Staaf for the original idea)

BUG=chrome-os-partner:6748
TEST=Compiled and run on Kaen and Stumpy

Change-Id: Icba93c3f36559e893a45bfb65a5cb711cb5b58a8
Reviewed-on: https://gerrit.chromium.org/gerrit/10180
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>