Prevent flashrom from detecting tegra3 as tegra2
Running flashrom on a tegra3 system was causing it to die in pretty
horrible ways.  We now prevent running on a tegra3 until flashrom
can be updated to support it.

NOTE that it may be non-trivial to support tegra3.  It looks like
tegra2_spi.c mucks directly with GPIO config / muxing, which is
pretty different in tegra3.

BUG=chromium-os:22481
TEST=Ran 'flashrom -V' on tegra2 and tegra3 systems.

On Tegra2 system:
  flashrom -V
  flashrom v0.9.3  :  :  on Linux 2.6.38.3 (armv7l), built with libpci 3.1.4, GCC 4.6.x-google 20110718 (prerelease), little endian
  Initializing internal programmer
  Detected NVIDIA Tegra 2.

On Tegra3 system:
  flashrom -V
  flashrom v0.9.3  :  :  on Linux 2.6.36.3-04239-g2ae7cf8 (armv7l), built with libpci 3.1.4, GCC 4.6.x-google 20110718 (prerelease), little endian
  Initializing internal programmer
  Processor detection/init failed.

Change-Id: I8e898baf95f02a0e32fa455ad311d39cd4e5a0ad
Reviewed-on: https://gerrit.chromium.org/gerrit/11136
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>