sync with upstream @ r1469
The most important change in this patchset is the introduction of
the opaque programmer model. This gives us a permanent (and less
hacky) solution to the decision of when to use hardware versus
software sequencing. Now it will select hardware sequencing if certain
opcodes are inaccessible due to lockdown or if >1 flash chip is
present (Same as the patch Louis committed earlier).

Note: The user can also select the mode with -p options:
-p internal:ich_spi_mode=hwseq
-p internal:ich_spi_mode=swseq
-p internal:ich_spi_mode=auto (default)

Also, flashrom.c was re-vamped to be more library-friendly, and a
lot of programmer-related functionality was moved from there to
the CLI code (cli_mfg in our case).

Other notable changes made in this commit to resolve merge conflicts:
cli_mfg.c
- Move programmer variable to cli_mfg.c (r1433 from upstream)
- Rename cli_mfg() to main().

Makefile
- Removed cli_classic.o from compilation. We use cli_mfg instead.
- Increase flashrom release number (r1394).

flashchips.c
- Removed hardware sequencing from list of flashchips. It has been
  replaced with the opaque programmer.

internal.c:
- hoisted internal programmer registration so that supported bus type checks
  in EC code would not break.
- Use ERROR_FATAL (r1426)
- Updated some print statements so they do not reference programmer. This
  is because the programmer variable is no longer exported from flashrom.c.

BUG=none
TEST=Test suite run on tested on zgb, alex, mario, stumpy, lumpy, kaen, aebl

Change-Id: Ied2838c16967c878e6c504f042c49b69d5010ccd