add common.sh and cleanup flashrom call sites
Introduce a common.sh file. This is intended to provide a central
location for some global variables and helper functions.

The first helper functions that are introduced are simple wrappers
around flashrom which run a command and suppress output. The intent
is to clean up several call sites where only the exit code is needed
to determine pass/failure.

Tests which need to examine the actual output of flashrom can still
use the old invocation.

BUG=none
TEST=tested on zgb

Change-Id: I2f9d97a16c355ce894f2d822ff72f2e1fa12e684
Reviewed-on: https://gerrit.chromium.org/gerrit/12172
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>