Lowercase non-global variable names
Earlier on, "global" meant outside the scope of an if-else or looping
construct.

Since we now have at least two scripts that share info with
the unit tests (do_tests.sh and common.sh), this patch changes that
definition to mean any variables that are not shared between scripts.
Any variable name that is isolated to a single script is now in
lowercase.

BUG=none
TEST=tested on zgb

Change-Id: I44650c5854f641fdac3be4e4ed1e6ef44d9208a7
Reviewed-on: https://gerrit.chromium.org/gerrit/12173
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>