Add check to ensure user-specified bus is available
This adds a small check to ensure that the bus specified by the
user is available.

This will change flashrom's behavior in some cases and fix a bug where
flashrom would succeed insteead of failing as expected. For our
current platforms, LPC support will only be made available when an EC
is successfully initialized and adds it to the list of supported
busses. This will cause operations with "-p internal:bus=lpc" to fail
(return non-zero) on platforms which do not have any devices connected
to LPC.

BUG=chrome-os-partner:7238
TEST=tested on zgb, mario, alex, lumpy, stumpy, kaen, and aebl.

Change-Id: I8eece3cfccda5bb70970fbba689f28a0d64ebf0d
Reviewed-on: https://gerrit.chromium.org/gerrit/13055
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>