GEC supports full chip update.
In short, the flashrom.c calls gec_need_2nd_pass() to decide if it needs
a 2nd pass of update, which happens after EC jumps to another copy and
overwrite the active image in the 1st pass. On the MP machine, only 1 pass
is needed because the firmware updater would specify the particular
section to update.

The EC would return ACCESS DENIED if the flashrom tries to write/erase
the active image (which the firmware copy EC is running on). Thus,
the gec.c can recognize which firmware copy is valid in new EC image
by comparing the denied address and new FMAP. Note that the ACCESS DENIED
error value is ignored just like the denied ME area (see ignore_error()).

gec_jump_copy() is smart. It can know which new firmware copy is available
to jump. It survives in the case that RO becomes smaller and RW A/B move
proceeding, which the old A conflicts to all of RO/A/B. In this case, the
gec_prepare() would first jump to RO. Then, doit() updates B. gec_jump_copy()
jumps from RO to B (rather than A because A was access denied by old RO in
the 1st pass), then finally update RO/A.

BUG=chrome-os-partner:8449,chrome-os-partner:chrome-os-partner:8603
TEST=Tested on the real board with the following combinations:
Initial active image: RO / A
Update: full update / -i RO / -i RW_A

Change-Id: I3cd97cb78927e673969714ff8cac482e9c3f84d2
Reviewed-on: https://gerrit.chromium.org/gerrit/18861
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Ready: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>
Tested-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>