Slightly change the spidev auto-detect logic.
The scan code is scanning /dev/spidevX.0 where X is from 0 to 9. The old
code stops while the first X makes stat() fails. This leads a bug if
the kernel spidev doesn't start from 0, for exmaple, bus 0 is skipped.

The new code is to keep trying even if stat() fails.

BUG=chrome-os-partner:9932
TEST=tested on daisy board. -r works.

Change-Id: Id7f4b7dc8bb15d7b8574eb830586d62daea4b88c
Reviewed-on: https://gerrit.chromium.org/gerrit/23486
Reviewed-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>
Tested-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>