CHERRY-PICK: search for spidev instead of m25p80
Kernel 3.4 does not seem to export the "m25p80" sysfs node,
only the "spidev" node.

This patch makes flashrom use the device matching "spidev" instead
of "m25p80" and with minor number 0 instead of 1. The major number
is what really matters.

(props to katierh for catching this in time for FSI!)

BUG=chrome-os-partner:9861,,chrome-os-partner:10420
TEST=Tested on Snow using kernels 3.3 and 3.4

Change-Id: If18b79a3e20fc7fcd6372d07b9ed6d749048080c
Reviewed-on: https://gerrit.chromium.org/gerrit/25092
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>