Do not call show_id() for now
This disables a sanity check that compares the vendor and mainboard
names from the ROM with the user-provided file for write operations.
The reason is because codenames which are in flux can cause failure.

BUG=chrome-os-partner:10414
TEST=Tested on Stumpy (notes below)

With this patch, the following text no longer appears on Stumpy
coreboot last image size (not ROM size) is 1048576 bytes.
Manufacturer: SAMSUNG
Mainboard ID: Stumpy

Change-Id: I759a2505d1944560dfec54157a9e9eaf28cdd56a
Reviewed-on: https://gerrit.chromium.org/gerrit/25633
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>