Apply new EC firmware ASAP after update.
The old code would stay in the RO code if the RO is not changed.
The new code would jump to new-updated copy (B is preferred than A).

BUG=chrome-os-partner:11156
Signed-off-by: Louis Yung-Chieh Lo <yjlou@chromium.org>
TEST=on link.
prepare EC images a and b almost identical except the RW part.
% flashrom -w a
% flashrom -w b
% ectool version
running on A  (new firmware)

Change-Id: I4d0a10a73d52997f49caa641a03a3a592c8f14a1
Reviewed-on: https://gerrit.chromium.org/gerrit/26914
Commit-Ready: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>
Reviewed-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>
Tested-by: Yung-Chieh Lo <yjlou%chromium.org@gtempaccount.com>