Make broken_timer flag configurable during compilation
This patch allows the broken_timer option to be set via a CONFIG
variable that can be set on the command-line (and hence as a USE
flag). This will allow us to set the option according to board.

Related ebuild change: https://gerrit.chromium.org/gerrit/#/c/29326/

BUG=chrome-os-partner:12292
TEST=Built for Daisy/Snow, regression tested for Alex and Stumpy

Change-Id: I2b7988ccc92d94d90d37d9516645671e8165d93a
Reviewed-on: https://gerrit.chromium.org/gerrit/29327
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31864