gec: add wait mechanism for in-progress commands
The EC can now indicate that a command has been accepted and is
in-progress. This was recently enabled for the flash write command:
https://gerrit.chromium.org/gerrit/#/c/30471/

This adds a retry mechanism which checks the status of the previous
command issued after an expected delay period for that command.

BRANCH=snow
BUG=none
TEST=tested on Snow using ToT EC image

Change-Id: I0c4522b351d84c13ff6873ceb85b1dce37b52efd
Reviewed-on: https://gerrit.chromium.org/gerrit/30801
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Yung-Chieh Lo <yjlou@chromium.org>