Remove tegra2_spi driver
This is generating warnings so may not work. Also it directly accesses
hardware from within Linux which is not great. It really should use spidev.

We don't need this at present - if it is needed in the future it can use
the spidev support. So I suggest we remove it.

BUG=none
BRANCH=none
TEST=emerge-daisy flashrom

Change-Id: Ia901cb90ed1e093f87b493324e2fb0c3556abc9c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43007