add read_flash() to do error handling for flash->read()
This adds read_flash() which acts as a wrapper around flash->read().
Its purpose is to make high-level decisions such as what to do in
case of an error.

When applicable, flash->read() call sites are updated to use this
wrapper in order to simplify the code.

BUG=chromium-os:38794
BRANCH=none
TEST=tested by doing partial read on Stumpy (with ME region locked),
     saw verbose output indicating ME region access error ignored

Change-Id: I4fe2a9a88d874783fbc1007f44434f7fafd08874
Reviewed-on: https://gerrit.chromium.org/gerrit/46207
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Yung-Chieh Lo <yjlou@chromium.org>