flashrom: Refactor search code to look for fdtmap and fmap simultaneously
If there is no FDTMAP available in the image, we end up doing an exhaustive
search for it before we try looking for an FMAP. It makes more sense to look
for both at the same time, since we don't expect to have more than one in
the same image.

Adjust the code to handle this.

BUG=chromium:224871
BRANCH=none
TEST=FEATURES=test sudo -E emerge flashrom

Change-Id: I88b1ad47063f14460db8b01923a0f741457e1c09
Reviewed-on: https://gerrit.chromium.org/gerrit/47214
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>