gec: Don't use the pass-through bus; use force mode of i2c
On snow we've got a special i2c 'pass through' bus that just creates
an i2c bus that mirrors all commands sent to that bus over to the same
bus that the EC is on.

There are a couple of properties of that bus:

1. Devices on this new bus aren't marked as BUSY by Linux, so you can
   open them without force mode.
2. This new bus can be found by its name, so if you want to find the
   i2c bus that contains the ec then you can search for this bus.

We are planning to remove the 'pass through' bus on snow in kernel 3.8
and redesign it a little on other devices that need to tunnel i2c
commands to other side of the ec.  In preparation, fix flashrom not to
need the i2c pass through bus.  This is done by:

1. Open up the i2c device using "force" mode.  This is really no worse
   than what we were doing before with the pass through but at least
   is straightforward about the fact that we're talking to the EC
   behind the kernel's back.
2. Find the proper bus by looking for the EC device, not the
   pass-through bus.

BUG=chrome-os-partner:18941
BRANCH=snow
TEST=On snow with kernel-next and UPSTREAM cros-ec can read EC with
flashrom -p internal:bus=lpc -r /tmp/ec.bin

Change-Id: I41d4b9cea0b90cb210be8d65f0b8cd1909f0da85
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49549
Reviewed-by: David Hendricks <dhendrix@chromium.org>