Skip Google EC if BUS_SPI given
This mirrors the check immediately above for the I2C-connected GEC. We need
this to be able to read from SPI flash with the 'internal' bus.

BUG=chromium:248372
BRANCH=none
TEST=manual
Run on pit:

flashrom -p internal:bus=spi -r foo2.bin

See that it now reads the SPI flash instead of the EC flash. There is no
EC traffic and the resulting file is 4MB in size.

Change-Id: Id6c1f9b59d94d947f57a839854b41deeba6acfd6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58704
Reviewed-by: David Hendricks <dhendrix@chromium.org>