Use region info to determine if erase/write violates access policy
Rather than get an error when we try to erase the wrong regions,
track the region that the EC is in using region information
provided by EC_CMD_FLASH_REGION_INFO.

This should also reduce transaction overhead substantially, which
may improve performance especially over I2C.

This is a combination of 60636 and 60319.

BUG=chrome-os-partner:20395
BRANCH=none
TEST=tested on link, snow, and spring by writing new image
with this CL, reading it back using ToT flashrom, and diffing the
images to see there is no corruption.

Change-Id: I15d62a337f62c0f05ff93344b072a7d66defb44d
Reviewed-on: https://gerrit.chromium.org/gerrit/60843
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>