Add support for auto-sizing of emulated ROMs
At present the flash size must be specified in command lines like:

   flashrom -p dummy:emulate=VARIABLE_SIZE,size=4M,image=xxx

In this case if there is an existing image, we can simply read that image
to figure out the size. So add a 'size=auto' option to make this more
convenient for the user.

BUG=chromium:256912
BRANCH=none
TEST=manual

sudo flashrom -p dummy:emulate=VARIABLE_SIZE,size=auto,image=/build/peach_pit \
/firmware/image-peach-pit.bin -x --ignore-lock

See that it no longer gives an error, but correctly decides on a 4MB image
size, and then completes the operation.

Change-Id: I4e738893ed0f106d75114de70eff1d599c7ed4e7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60860
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>