Read only as much ROM content as needed before writing
When we are using --fast-verify we make the assumption that the
full verification is not needed. Therefore it does not make sense
to read the entire flash before writing.

This makes it so that we only read the bytes required to verify
content we are writing when --fast-verify is used.

BUG=none
BRANCH=none
TEST=See below

-  Test partial write using --fast-verify:
   flashrom -p host -r foo.bin
   dd if=/dev/urandom of=random.bin bs=1 count=16384
   flashrom -p host -i RW_ELOG:random.bin -w foo.bin --fast-verify -V
   flashrom -p host -i RW_ELOG:random.bin -v foo.bin --fast-verify -V

   # verify against full file with and without --fast-verify
   dd if=random.bin of=foo.bin bs=1 seek=$((0x3f8000)) conv=notrunc
   flashrom -p host -i RW_ELOG --fast-verify -v foo.bin -V
   flashrom -p host -i RW_ELOG -v foo.bin -V

Change-Id: Iea5f3cc140e1b30644d4452daa64709eaf03b527
Reviewed-on: https://gerrit.chromium.org/gerrit/63118
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>