Read entire flash if VERIFY_PARTIAL is not used
In 52ddff, the rules for obtaining a reference image were modified
and centered around verify_it. However, we neglected the VERIFY_OFF
case which occurs when --noverify is used.

This patch will make it so that the VERIFY_PARTIAL (--fast-verify)
case is checked first. If not, then the entire chip is read. This
makes it so that common blocks are skipped during erase/write when
--noverify is used.

BUG=none
BRANCH=none
TEST=flashed an image using -noverify, unchanged blocks are no longer
erased and re-written.

Change-Id: I44539c5702f9bcc22ea28644f7cb30f6c531b647
Reviewed-on: https://gerrit.chromium.org/gerrit/64410
Reviewed-by: Yung-Chieh Lo <yjlou@chromium.org>
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>