Fix use of uninitialized local variable in gec_jump_copy()
The code in get_jump_copy() uses a local structure variable, which is
never initialized - a leftover of a previous fix/clean up. This causes
interesting side effects depending on the stack contents at the moment
the function is invoked. In particular enabling verbose mode makes
programming succeed, while it always fails when verbose debug is
disabled.

A bigger question is why this package tolerates warnings when
building, this will have to be fixed separately.

BRANCH=r30
BUG=chromium:295082
TEST=manual

   . running the following command on pit does not fail anymore:

   $ ./flashrom -p ec -w ec.25.bin && ./flashrom -p ec -w ec.52.bin

   before this fix it would always fail on the very first attempt.

Change-Id: If77b0c166045adeee0779572260776eec832403f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170029
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: David Hendrix <dhendrix@chromium.org>