Do not tolerate warning when building flashrom
Compilation warnings should never be tolerated, while some of them a
re harmless and annoying, sometimes they are an indication of a
serious problem (as in the bug referred below). Do not tolerate
warnings while building flashrom using default invocation.

If someone is missing compiling with warnings tolerated - allow to do
it invoking make with NOWARNERROR=yes.

BRANCH=None
BUG=chromium:295082
TEST=manual
   . successfully emerged flashrom for hos, link and pit

Change-Id: Ia6e0b65bff7bfa1539c75ca5e21cfc402069bf92
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170073
Reviewed-by: David Hendrix <dhendrix@chromium.org>