CrOS EC: Fallback to I2C if /dev/cros_ec is unreliable
This adds a test to see if we can communicate with the EC via
/dev/cros_ec interface. cros_ec is generally preferred, especially
for newer platforms, but may fail on older ones with kernel 3.4.

TODO: The "hello" test for CrOS EC is already used in multiple places
and this adds yet another. A follow-up CL will be sent to remove
duplicate code.

BUG=chromium:314765
BRANCH=none
TEST=ran "flashrom -p ec -r /tmp/foo.bin -V" on Snow with 3.4 kernel,
saw it fall back to raw I2C when cros_ec failed.

Change-Id: I344bbef058f7d395e74406c619539890b25d61d2
Reviewed-on: https://chromium-review.googlesource.com/175672
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Yung-chieh Lo <yjlou@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>