linux_spi: search for spi:spidev first
This reverses the order in which SPI modaliases are searched. It
turns out that the search can be surprisingly slow even though
only a few system calls are used, and seems to get much worse
on every iteration. So we'll try to reduce the number of iterations
by searching for what we expect to find first.

Another approach would be to make scanft() take an array of
proverbial needles to compare with every matching file so we would
only parse sysfs once. But that is a more ambitious approach that
will require a lot more work and testing.

BUG=chrome-os-partner:24007
BRANCH=none
TEST=Tested on pit, "flashrom -p host -i RW_ELOG:/tmp/elog.bin -r"
takes ~100ms instead of ~1.5s.

Change-Id: Id58d327af825141913f750e74a534681f10d0dd6
Reviewed-on: https://chromium-review.googlesource.com/176556
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>