GD25Q32: Mark as tested, enable write-protection
This marks GD25Q32 as tested and adds the .wp member to the
flashchip struct.

Since there are some discrepancies between the write-protect
ranges between W25Q32 and GD25Q32, and because it uses 6 bits for
for block-protect range information in a manner that doesn't fit
comfortably in the w25q32 mold, this chip will be one of the first
to use the more generic write-protect infrastructure.

To make things more exciting, GD25LQ32 has a "CMP" bit in status
register 2 which changes the granularity from blocks to sectors.
This could get weird if changing the CMP bit is required to attain
the desired level of block protection.

BUG=None
BRANCH=None
TEST=wrote ROM on a GD25Q32 chip via Servo

Change-Id: Ib124b61de598000b6780b866e090d0265da82483
Reviewed-on: https://chromium-review.googlesource.com/210398
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>