Add partial support for S25FL256S.
This patch adds support for a new variety of Spansion chips. There is
a pressing need to be able to read and write the bottom half of the
S25FL256S Spansion chips, this patch introduces partial support: these
chips are recognized, but treated as 128Mb devices instead of 256Mb.

Currently supported variety (25FSxxxx) while named very similarly is
logically different from 25FLxxxx:

    - FL chips do not support RDAR/WRAR commands;

    - some FL chips have overlayed mixed sized sector architecture and
      require 4 byte address for the sector erase command to make sure
      the sector of the desired size is being erased

Another problem is that corresponding S25FSxxxx and S25FLxxxx chips
return the same device ID in the first three bytes of the RDID command
output. To be able to tell between these two varieties one needs to
look at five bytes out of the first six returned by the RDID command.

The structure of the RDID output is as follows:

offset   value              meaning
  00h     01h      Manufacturer ID for Spansion
  01h     20h           128 Mb capacity
  01h     02h           256 Mb capacity
  02h     18h           128 Mb capacity
  02h     19h           256 Mb capacity
  03h     4Dh       Full size of the RDID output (ignored)
  04h     00h       FS: 256-kB physical sectors
  04h     01h       FS: 64-kB physical sectors
  04h     00h       FL: 256-kB physical sectors
  04h     01h       FL: Mix of 64-kB and 4KB overlayed physical sectors
  05h     80h       FL family
  05h     81h       FS family

This patch introduces the new probing function, which looks at the
five out of the first six bytes to tell which exact chip is being
probed.

The AR register is not defined in the FL variety, and there is no way
to change the sector size, the user is supposed to use the chip
specific sector erase command (DCh) which requires 4 byte address.

BRANCH=storm
BUG=chrome-os-partner:32522
TEST=was able to program both storm (using "S25FS128S Small Sectors")
     and whirlwind (using "S25FL256S Small Sectors") devices.

Change-Id: I2a766602735f61c068b028dbfa375d4bdb004c9e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/224904
Reviewed-by: Bryan Freed <bfreed@chromium.org>