writeprotect: Simplify status register write usage
We have multiple implementations of spi_write_status_register() which
appear to have been written at approximately the same time. The
generic one has an advantage in that it polls the write-in-progress
bit to see if WRSR has completed instead of relying on a single hard-
coded delay.

Up until now, the chips we've been working with so far have worked
with the old hard-coded delay of 100ms. However, the S25FS128
can take up to 750ms to write to nonvolatile registers. So this
is the impetus for switching over to using the generic function
instead of the nearly identical version we have in writeprotect.c.

For now, the W25Q stuff is left alone.

BUG=chrome-os-partner:34430
BRANCH=none
TEST=tested on storm using --wp-range

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: Ic0d41089cc8f4a4fd760aca34e705a42d7520ddf
Reviewed-on: https://chromium-review.googlesource.com/234910
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>