Align partial reads to eraseblock size for partial writes
If a partial write operation is specified on a region that is not
aligned to an eraseable block size, the data outside that region
will be lost since the partial read function only backed-up the
data which resided in the region specified by the user and the
block erase command will wipe out the entire block.

This patch fixes handle_partial_read() so that it takes eraseable
block size into account. For now the heuristic is crude and assumes
that the smallest eraseable block size will be used. With this,
we can adjust the start and length to take data within the entire
block into account when reading old content from the flash chip.

This was likely overlooked since we usually re-write block-aligned
regions. However, some large flash chips have large block sizes which
can contain multiple regions.

BUG=chromium:428475, chrome-os-partner:34430
BRANCH=none
TEST=set_gbb_flags.sh on storm no longer wipes out the adjacent
FMAP region on the same block

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I3c06744b06650b237ea6841eee7d499fd764ee3b
Reviewed-on: https://chromium-review.googlesource.com/240176