Don't verify if ROM content hasn't changed
If no changes are made to the content of the firmware ROM, for example
if the user tells flashrom to write an image that is the same as
what's currently on the ROM, then the final verification step becomes
redundant and wastes time.

This patch adds a variable that will be set if any erase or write
operation is to be performed. Verification will still occur if the
user passes the "-v" flag or if an erase/write operation fails for
any reason (even if it's on the programmer side and changes have not
been committed to the chip).

BUG=none
BRANCH=none
TEST=wrote an image to veyron_speedy using "-w" and verification
occurred, wrote the same image again and verification was skipped.
Also tested partial write the same way using "--fast-verify".

Change-Id: I1bef21ef71669a7858b9654f987338e04c881a28
Reviewed-on: https://chromium-review.googlesource.com/263978
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>